+++ SampleProjects/SharedLibrary/src/SharedLibrary.cpp
@@ -1,0 +1,5 @@
+#include "SharedLibrary.h"
+
+int main() {
+  return foo; // 'foo' was not declared in this scope
+}